Revert "Modular run taks (#33)"
This reverts commit a2caa53a0b9b09c5da855f8e257e44ee837ef297.
