--- rewardgym/runner/__init__.py
@@ -1,5 +1,0 @@
-from .file_utils import make_bids_name
-from .psychopy_runner import pspy_run_task
-from .psychopy_utils import pspy_set_up_experiment
-
-__all__ = ["pspy_run_task", "make_bids_name", "pspy_set_up_experiment"]
