+++ rewardgym/tests/test_psychopy_stubs_smoke.py
@@ -1,9 +1,9 @@
 import pytest
 
 from rewardgym import get_env
+from rewardgym.psychopy_core import run_task
 from rewardgym.psychopy_render import SimulationLogger
 from rewardgym.psychopy_render.psychopy_stubs import Clock, Window
-from rewardgym.runner import pspy_run_task
 
 
 def simulate_task(envname):
@@ -26,7 +26,7 @@ def simulate_task(envname):
     env = get_env(task, render_backend="psychopy-simulate")
     env.setup(logger=Logger, window=win, expose_last_stim=True)
 
-    pspy_run_task(env=env, win=win, logger=Logger)
+    run_task(env=env, win=win, logger=Logger)
 
     return Logger.close()
